+++ presentation/src/main/java/com/example/miso/ui/main/component/main/LogoutButton.kt
@@ -1,4 +1,4 @@
-package com.example.miso.ui.main.component
+package com.example.miso.ui.main.component.main
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -14,24 +14,24 @@ import androidx.compose.ui.unit.dp
 import com.example.miso.R
 
 @Composable
-fun SearchButton(
+fun LogoutButton(
     onClick: () -> Unit
 ) {
     Image(
-        painter = painterResource(id = R.drawable.ic_search),
+        painter = painterResource(id = R.drawable.ic_logout),
         contentDescription = "Logout Icon",
         contentScale = ContentScale.FillBounds,
         modifier = Modifier
-            .size(width = 24.dp, height = 24.dp)
+            .size(width = 14.dp, height = 22.dp)
             .clickable(
                 interactionSource = remember { MutableInteractionSource() },
                 indication = null
-            ) { onClick }
+            ) { onClick() }
     )
 }
 
 @Composable
 @Preview(showBackground = true)
-fun SearchButtonPreView() {
-    SearchButton({})
+fun LogoutButtonPreView() {
+    LogoutButton({})
 }
